requirements.txt
streamlit
websocket-client
pandas
numpy
statsmodels
plotly
